[refactor][feature] dynamic task def injections
Changed all task definitions to look like each other - not wrapped in a module
This is so that readers of src wont get confused when they just want to write more tasks.

The DSL object treats the default task dir the same way it would for an injected path.
i.e.: it's the same logic through & through, the only difference is that the default task
      path gets loaded on initialization of the object

Also too the time to remove the consts in the tasks